+++ common/src/main/java/com/kryeit/content/jar_of_tips/FallingBlockEntityHelper.java
@@ -1,0 +1,8 @@
+package com.kryeit.content.jar_of_tips;
+
+import net.minecraft.world.level.block.state.BlockState;
+
+public interface FallingBlockEntityHelper {
+    BlockState missions$getBlockState();
+    void missions$setBlockState(BlockState state);
+}