+++ common/src/main/java/com/kryeit/registry/ModBlocks.java
@@ -2,6 +2,8 @@ package com.kryeit.registry;
 
 import com.kryeit.Missions;
 import com.kryeit.content.exchanger.MechanicalExchangerBlock;
+import com.kryeit.content.jar_of_tips.JarOfTipsBlock;
+import com.kryeit.content.jar_of_tips.JarOfTipsItem;
 import com.simibubi.create.content.kinetics.BlockStressDefaults;
 import com.simibubi.create.foundation.data.BlockStateGen;
 import com.simibubi.create.foundation.data.SharedProperties;
@@ -12,7 +14,6 @@ import net.minecraft.world.level.block.state.BlockBehaviour;
 
 import static com.kryeit.registry.ModCreativeTabs.useBaseTab;
 import static com.simibubi.create.foundation.data.ModelGen.customItemModel;
-import static com.simibubi.create.foundation.data.TagGen.axeOrPickaxe;
 
 public class ModBlocks {
 
@@ -26,11 +27,21 @@ public class ModBlocks {
             .properties(BlockBehaviour.Properties::noOcclusion)
             .addLayer(() -> RenderType::translucent)
             .transform(BlockStressDefaults.setImpact(32.0))
-            .transform(axeOrPickaxe())
             .blockstate(BlockStateGen.horizontalBlockProvider(true))
             .item()
             .properties(p -> p.rarity(Rarity.UNCOMMON))
             .transform(customItemModel())
+            .register();
+
+    public static final BlockEntry<JarOfTipsBlock> JAR_OF_TIPS = Missions.registrate()
+            .block("jar_of_tips", JarOfTipsBlock::new)
+            .properties(p -> p
+                    .strength(0.1F)
+                    .noOcclusion())
+            .addLayer(() -> RenderType::translucent)
+            .blockstate(BlockStateGen.horizontalBlockProvider(true))
+            .item(JarOfTipsItem::new)
+            .build()
             .register();
 
     public static void register() {}